Hop: merged parts into main hop_runner
