+++ tool/hop_runner.dart
@@ -6,10 +6,6 @@ import "dart:io";
 import 'package:hop/hop.dart';
 import 'package:hop/hop_tasks.dart' hide createAnalyzerTask;
 
-part 'docgen.dart';
-part 'utils.dart';
-part 'analyze.dart';
-
 var files = ["lib/files.dart", "lib/html.dart", "lib/io.dart"];
 
 void main(List<String> args) {
@@ -19,4 +15,71 @@ void main(List<String> args) {
   addTask("bench", createBenchTask());
   addChainedTask("check", ["analyze"]);
   runHop(args);
+}
+
+Task createAnalyzerTask(List<String> files, [List<String> extra_args]) {
+  var args = [];
+  args.addAll(files);
+  if (extra_args != null) {
+    args.addAll(extra_args);
+  }
+  return createProcessTask("dartanalyzer", args: args, description: "Statically Analyze Code");
+}
+
+/* Custom DocGen Task for Flexibility */
+Task createDocGenTask(String path, {compile: false, Iterable<String> excludes: null, include_sdk: true, include_deps: false, out_dir: "docs", verbose: false}) {
+  return new Task((TaskContext context) {
+    var args = [];
+
+    if (verbose) {
+      args.add("--verbose");
+    }
+
+    if (excludes != null) {
+      for (String exclude in excludes) {
+        context.fine("Excluding Library: ${exclude}");
+        args.add("--exclude-lib=${exclude}");
+      }
+    }
+
+    if (compile) {
+      args.add("--compile");
+    }
+
+    if (include_sdk) {
+      args.add("--include-sdk");
+    } else {
+      args.add("--no-include-sdk");
+    }
+
+    if (include_deps) {
+      args.add("--include-dependent-packages");
+    } else {
+      args.add("--no-include-dependent-packages");
+    }
+
+    args.add("--out=${out_dir}");
+
+    args.addAll(context.arguments.rest);
+
+    args.add(path);
+
+    context.fine("using argments: ${args}");
+
+    return Process.start("docgen", args).then((process) {
+      return inheritIO(process);
+    }).then((code) {
+      if (code != 0) {
+        context.fail("docgen exited with the status code ${code}");
+      }
+    });
+  }, description: "Generates Documentation");
+}
+
+Future<int> inheritIO(Process process) {
+  process.stdin.addStream(stdin);
+  stdout.addStream(process.stdout);
+  stderr.addStream(process.stderr);
+
+  return process.exitCode;
 }
